SSCS-5994 - Move code to one handler
