--- src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/actionfurtherevidence/ActionFurtherEvidenceReceivedHandler.java
@@ -1,50 +1,0 @@
-package uk.gov.hmcts.reform.sscs.ccd.presubmit.actionfurtherevidence;
-
-import static java.util.Objects.requireNonNull;
-import static uk.gov.hmcts.reform.sscs.ccd.presubmit.actionfurtherevidence.FurtherEvidenceActionDynamicListItems.*;
-
-import org.apache.commons.lang.StringUtils;
-import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
-import uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType;
-import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
-import uk.gov.hmcts.reform.sscs.ccd.domain.*;
-import uk.gov.hmcts.reform.sscs.ccd.presubmit.PreSubmitCallbackHandler;
-
-
-@Service
-public class ActionFurtherEvidenceReceivedHandler implements PreSubmitCallbackHandler<SscsCaseData> {
-
-    private static final String FURTHER_EVIDENCE_RECEIVED = "furtherEvidenceReceived";
-
-    public boolean canHandle(CallbackType callbackType, Callback<SscsCaseData> callback) {
-        requireNonNull(callback, "callback must not be null");
-        requireNonNull(callbackType, "callbacktype must not be null");
-
-        return callbackType.equals(CallbackType.ABOUT_TO_SUBMIT)
-            && callback.getEvent() == EventType.ACTION_FURTHER_EVIDENCE
-            && isIssueFurtherEvidenceToAllParties(callback.getCaseDetails().getCaseData().getFurtherEvidenceAction());
-
-    }
-
-    private boolean isIssueFurtherEvidenceToAllParties(DynamicList furtherEvidenceActionList) {
-        if (furtherEvidenceActionList != null && furtherEvidenceActionList.getValue() != null
-                && StringUtils.isNotBlank(furtherEvidenceActionList.getValue().getCode())) {
-            return furtherEvidenceActionList.getValue().getCode().equals(ISSUE_FURTHER_EVIDENCE.getCode());
-        }
-        return false;
-    }
-
-    public PreSubmitCallbackResponse<SscsCaseData> handle(CallbackType callbackType, Callback<SscsCaseData> callback) {
-        if (!canHandle(callbackType, callback)) {
-            throw new IllegalStateException("Cannot handle callback");
-        }
-
-        final SscsCaseData caseData = callback.getCaseDetails().getCaseData();
-        caseData.setDwpFurtherEvidenceStates(FURTHER_EVIDENCE_RECEIVED);
-
-        return new PreSubmitCallbackResponse<>(caseData);
-    }
-
-
-}
