+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/actionfurtherevidence/HandleEvidenceEventHandlerTest.java
@@ -8,9 +8,12 @@ import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 import junitparams.converters.Nullable;
@@ -20,16 +23,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
-import uk.gov.hmcts.reform.sscs.ccd.domain.CaseDetails;
-import uk.gov.hmcts.reform.sscs.ccd.domain.DocumentLink;
-import uk.gov.hmcts.reform.sscs.ccd.domain.DynamicList;
-import uk.gov.hmcts.reform.sscs.ccd.domain.DynamicListItem;
-import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.ScannedDocument;
-import uk.gov.hmcts.reform.sscs.ccd.domain.ScannedDocumentDetails;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsDocument;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsDocumentDetails;
+import uk.gov.hmcts.reform.sscs.ccd.domain.*;
 
 @RunWith(JUnitParamsRunner.class)
 public class HandleEvidenceEventHandlerTest {
@@ -226,6 +220,29 @@ public class HandleEvidenceEventHandlerTest {
         for (String error : response.getErrors()) {
             assertEquals("No further evidence to process", error);
         }
+    }
+
+    private Callback<SscsCaseData> buildCallback() {
+        DynamicList dynamicList = new DynamicList(new DynamicListItem("issueFurtherEvidence", "label"),
+                Collections.singletonList(new DynamicListItem("issueFurtherEvidence", "label")));
+        SscsCaseData sscsCaseData = SscsCaseData.builder()
+                .originalSender(dynamicList)
+                .furtherEvidenceAction(dynamicList)
+                .scannedDocuments(Collections.singletonList(ScannedDocument.builder().build()))
+                .build();
+        CaseDetails<SscsCaseData> caseDetails = new CaseDetails<>(123L, "sscs",
+                State.INTERLOCUTORY_REVIEW_STATE, sscsCaseData, LocalDateTime.now());
+        return new Callback<>(caseDetails, Optional.empty(), EventType.ACTION_FURTHER_EVIDENCE);
+    }
+
+    @Test
+    public void givenIssueFurtherEvidence_shouldUpdateDwpFurtherEvidenceStates() {
+        Callback<SscsCaseData> callback = buildCallback(
+        );
+
+        PreSubmitCallbackResponse<SscsCaseData> updated = handleEvidenceEventHandler.handle(ABOUT_TO_SUBMIT, callback);
+
+        assertEquals("furtherEvidenceReceived", updated.getData().getDwpFurtherEvidenceStates());
     }
 
 }
